+++ build.zig
@@ -54,11 +54,6 @@ pub fn build(b: *std.Build) void {
         "compare-refs",
         "Comma-separated list of git refs to compare benchmarks against (for bench-compare step)",
     ) orelse "v0.1.0";
-    const min_index_capacity = b.option(
-        u32,
-        "min-index-capacity",
-        "Minimum capacity to initially allocate for the TOML hash table indices",
-    ) orelse 16;
     const random_bench = b.option(
         bool,
         "random-bench",
@@ -70,15 +65,11 @@ pub fn build(b: *std.Build) void {
         "Timeout value to pass to 'toml-test' runs. The value is passed to 'toml-test' as is. Default is '1s'",
     ) orelse "5s";
 
-    const options = b.addOptions();
-    options.addOption(u32, "min_index_capacity", min_index_capacity);
-
     const toml_mod = b.createModule(.{
         .root_source_file = b.path("src/root.zig"),
         .target = target,
         .optimize = optimize,
     });
-    toml_mod.addOptions("build_options", options);
 
     // Add the library to the package's module set
     b.modules.put("toml", toml_mod) catch @panic("OOM");